+++ packages/code-input/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 2.3.1
+
+### Patch Changes
+
+### [#418](https://github.com/core-ds/core-components/pull/418)
+
+-   Исправлена проблема с default-импортом в cjs форматах
+
+## 2.3.0
+
+### Minor Changes
+
+### [#407](https://github.com/core-ds/core-components/pull/407)
+
+-   В CodeInput добавлен новый проп onErrorAnimationEnd
+-   В Confirmation теперь сбрасывается состоянии ошибки при включенном пропе clearCodeOnError<br />
+-   В Confirmation исправлена ошибка, из-за которой не показывался текст из пропа texts.blockSmsRetry<br />
+
 ## 2.2.0
 
 ### Minor Changes