+++ packages/select-with-tags/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 7.0.10
+
+### Patch Changes
+
+### [#418](https://github.com/core-ds/core-components/pull/418)
+
+-   Исправлена проблема с default-импортом в cjs форматах
+
+-   Обновлены зависимости
+    -   form-control@9.0.2
+    -   select@13.2.3
+    -   tag@5.0.3
+
+## 7.0.9
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@13.2.2
+
 ## 7.0.8
 
 ### Patch Changes